Refactor: Extract GpuWorkQueue to src/analysis/gpu/queue.rs (#274)
Move the `GpuWorkQueue` struct and thread management code from
`implementation.rs` (~500 lines) to a dedicated `gpu/queue.rs` module
as part of the ongoing implementation.rs monolith refactoring.

## Changes

- Create `src/analysis/gpu/queue.rs` with:
  - `GpuWorkRequest` enum (all GPU operation types)
  - `GpuWorkQueue` struct (centralised GPU work queue)
  - Thread creation, main processing loop, graceful shutdown
  - `impl Drop` with timeout protection
  - `impl GpuEvaluator` for polymorphic GPU evaluation
  - Unit tests for the queue module

- Update `src/analysis/gpu/mod.rs`:
  - Add `pub mod queue` declaration
  - Re-export `GpuWorkQueue` for backwards compatibility
  - Update module documentation

- Update `src/analysis/implementation.rs`:
  - Remove ~500 lines of GpuWorkQueue code
  - Update imports to use new queue module
  - Remove unused imports

## Success Criteria

- [x] GpuWorkQueue moved to src/analysis/gpu/queue.rs
- [x] Thread naming preserved for diagnostics
- [x] Graceful shutdown behaviour preserved
- [x] All existing tests pass (./quality.sh)
- [x] No public API changes

Parent issue: #185 (Complete refactoring of implementation.rs monolith)
