+++ docs/pr-summary-274.md
@@ -1,0 +1,62 @@
+## Summary
+
+Refactored `GpuWorkQueue` struct and thread management code from `implementation.rs` to a dedicated `src/analysis/gpu/queue.rs` module as part of the ongoing implementation.rs monolith refactoring (Issue #185).
+
+### Changes
+
+1. **Created `src/analysis/gpu/queue.rs`** (~530 lines):
+   - `GpuWorkRequest` enum - All GPU operation types (HelpfulBatch, HarmfulBatch, ReluEval, ActivationEval, Shutdown)
+   - `GpuWorkQueue` struct - Centralised GPU work queue for thread-based GPU work distribution
+   - `impl GpuWorkQueue` - Thread creation with proper naming, main processing loop, graceful shutdown
+   - `impl Drop for GpuWorkQueue` - Clean resource release with timeout protection
+   - `impl GpuEvaluator for GpuWorkQueue` - Trait implementation for polymorphic GPU evaluation
+   - `GPU_SHUTDOWN_TIMEOUT_SECS` constant - 10 second timeout for graceful shutdown
+
+2. **Updated `src/analysis/gpu/mod.rs`**:
+   - Added `pub mod queue` declaration
+   - Added `pub use queue::GpuWorkQueue` re-export for backwards compatibility
+   - Updated module documentation to reflect refactoring progress
+
+3. **Updated `src/analysis/implementation.rs`**:
+   - Removed ~500 lines of `GpuWorkQueue` code
+   - Updated imports to use the new queue module
+   - Removed unused imports (crossbeam_channel types, Duration, JoinHandle)
+   - Added comments documenting what was moved and where
+
+### Architecture
+
+The `GpuWorkQueue` was introduced in v0.1.151 to solve a critical performance issue:
+- Previously each parallel focus neuron thread created its own GPU device (~100ms overhead)
+- Now a single shared queue owns the GPU thread
+- Operations are serialised through the queue, reducing device creation overhead
+- `Arc<GpuWorkQueue>` implements `GpuEvaluator` for polymorphic use
+
+## Evidence
+
+Unable to generate screenshot: This is a CLI-only library with no visual interface. The change is a pure code refactoring with no functional changes to the public API.
+
+## Test Plan
+
+- All existing tests pass (`./quality.sh` completes successfully)
+- Added unit tests in `src/analysis/gpu/queue.rs`:
+  - `test_queue_module_exports_are_accessible` - Verifies GpuWorkQueue struct is accessible
+  - `test_gpu_work_queue_implements_evaluator` - Verifies GpuEvaluator trait implementation
+  - `test_empty_helpful_batch_returns_empty_vec` - Tests early return for empty batches
+  - `test_empty_harmful_batch_returns_empty_vec` - Tests early return for empty batches
+  - `test_empty_relu_samples_returns_defaults` - Tests early return for empty samples
+  - `test_empty_activation_samples_returns_defaults` - Tests early return for empty samples
+  - `test_shutdown_timeout_is_reasonable` - Compile-time verification of timeout constant
+  - `test_gpu_work_request_variants_constructible` - Verifies all enum variants can be constructed
+- Integration tests in `tests/gpu_work_queue.rs` continue to pass:
+  - `synapse_analysis_works_with_gpu_work_queue`
+  - `neuron_analysis_works_with_gpu_work_queue`
+  - `harmful_synapse_detection_works_with_queue`
+  - `multiple_focus_neurons_work_with_shared_queue`
+
+## Success Criteria (from Issue #274)
+
+- [x] `GpuWorkQueue` moved to `src/analysis/gpu/queue.rs`
+- [x] Thread naming preserved for diagnostics
+- [x] Graceful shutdown behaviour preserved
+- [x] All existing tests pass (`./quality.sh`)
+- [x] No public API changes
